pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "blender-doc-tool"
version = "0.1.0"
description = "Comprehensive documentation tool for Blender projects"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [
    {name = "Blender Doc Team"}
]
keywords = ["blender", "documentation", "asset-management", "digraph", "pdf"]

dependencies = [
    "bpy>=4.0.0",
    "networkx>=3.0",
    "reportlab>=4.0.0",
    "Pillow>=10.0.0",
    "pandas>=2.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "flake8>=6.0.0",
    "mypy>=1.0.0",
]

[project.scripts]
blender-doc = "blender_doc.cli:main"

[tool.setuptools]
packages = ["blender_doc"]
package-dir = {"" = "src"}

[tool.black]
line-length = 100

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
